+++ mechanics-module/src/main/java/marowak/dev/request/PlayerMotionRequest.java
@@ -2,7 +2,7 @@ package marowak.dev.request;
 
 public record PlayerMotionRequest(
         boolean isUpdate,
-        Integer speed,
-        Integer angle
+        float speed,
+        int angle
 ) {
 }
